Track and sort by received-time for request batches.

+++ manager/src/main/scala/org/broadinstitute/transporter/web/Page.scala
@@ -7,9 +7,9 @@ import io.circe.derivation.{deriveDecoder, deriveEncoder}
   * Generic API envelope for a page of query results.
   *
   * @param items query results to return to the client
-  * @param total the number of elements included in `items`
+  * @param total the total number of items stored across all pages
   */
-case class Page[I](items: List[I], total: Int)
+case class Page[I](items: List[I], total: Long)
 
 object Page {
   implicit def decoder[I: Decoder]: Decoder[Page[I]] = deriveDecoder
